requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
sqlalchemy==2.0.23
python-dateutil==2.8.2
certifi==2024.2.2
pytest==7.4.3 